+++ packages/components/src/components/table/table.e2e.ts
@@ -1,4 +1,4 @@
-import { newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
+import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it } from "vitest";
 import { html } from "../../../support/formatting";
 
@@ -14,6 +14,23 @@ import type { TableHeader } from "../table-header/table-header";
 import type { TableCell } from "../table-cell/table-cell";
 import { TableRow } from "../table-row/table-row";
 import { SLOTS } from "./resources";
+
+async function getFocusedTableCellClassList(
+  page: E2EPage,
+  rowId: string,
+  cellSelector: "td" | "th",
+): Promise<string[]> {
+  return page.$eval(
+    `#${rowId}`,
+    (el, cellSelector) => {
+      const activeElement = el.shadowRoot!.activeElement as HTMLElement;
+      const cell = activeElement.shadowRoot!.querySelector<HTMLElement>(cellSelector)!;
+
+      return Array.from(cell.classList);
+    },
+    cellSelector,
+  );
+}
 
 describe("selection modes", () => {
   it("selection mode single allows one or no rows to be selected", async () => {
@@ -130,7 +147,6 @@ describe("selection modes", () => {
     expect(await element.getProperty("selectedItems")).toEqual([]);
     await selectedItemAsserter([]);
   });
-
   it("selection mode multiple allows one, multiple, or no rows to be selected", async () => {
     const page = await newE2EPage();
     await page.setContent(
@@ -1736,11 +1752,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(`#${rowHead.id}`, async (el) => {
-        return el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList;
-      }),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1752,27 +1766,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1788,12 +1796,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
   });
 
   it("navigates correctly when number column present", async () => {
@@ -1830,12 +1835,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1847,27 +1849,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1883,12 +1879,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when number and selection column present numbered", async () => {
@@ -1925,22 +1918,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1948,33 +1935,25 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1986,12 +1965,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when pagination present and selection and number and first page displayed", async () => {
@@ -2032,22 +2008,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2055,27 +2025,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2089,23 +2053,17 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when pagination present, and selection and number and navigation to two other pages occurs", async () => {
@@ -2149,21 +2107,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(`#${rowHead.id}`, (el) => {
-        return el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList;
-      }),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2171,45 +2124,39 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
@@ -2219,23 +2166,17 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.$eval(
       "calcite-table",
@@ -2258,45 +2199,39 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row4.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row4.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row4.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row4.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
@@ -2306,12 +2241,9 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when number and selection column present numbered and interaction-mode static - only focusing selection cells", async () => {
@@ -2349,58 +2281,44 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual([HEADER_CSS.selectionCell]);
   });
 
   it("updates table-row's selected property correctly when calciteTabRowSelect event is emitted", async () => {
@@ -2423,8 +2341,8 @@ describe("keyboard navigation", () => {
     );
 
     async function selectRow(rowSelector: string): Promise<void> {
-      await page.$eval(rowSelector + " >>> calcite-table-cell:first-child", (el) => {
-        (el as TableCell["el"]).click();
+      await page.$eval(rowSelector + " >>> calcite-table-cell:first-child", (el: TableCell["el"]) => {
+        el.click();
       });
       await page.waitForChanges();
     }
@@ -2455,5 +2373,212 @@ describe("keyboard navigation", () => {
     const rowDeselected = await propValueAsserter(false);
     await selectRow(rowSelector);
     await expect(rowDeselected()).resolves.toBe(undefined);
+  });
+
+  describe("setting current page", () => {
+    it("starts on page in range and programmatically changes page", async () => {
+      const page = await newE2EPage();
+      await page.setContent(
+        html`<calcite-table id="calcite-table" caption="Simple table" page-size="6" bordered current-page="2">
+          <calcite-table-row slot="table-header">
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 1</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 2</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 3</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 4</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 5</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row selected>
+            <calcite-table-cell>cell content 6</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row selected>
+            <calcite-table-cell>cell content 7</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 8</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 9 </calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 10</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 11</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 12</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 13</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+        </calcite-table>`,
+      );
+
+      const table = await page.find("calcite-table");
+
+      expect(await table.getProperty("currentPage")).toBe(2);
+
+      table.setProperty("currentPage", 3);
+      await page.waitForChanges();
+
+      expect(await table.getProperty("currentPage")).toBe(3);
+
+      const chevron = await page.find(`calcite-table >>> calcite-pagination >>> .${PAGINATION_CSS.chevron}`);
+      await chevron.click();
+
+      expect(await table.getProperty("currentPage")).toBe(2);
+
+      const numberLink = await page.find(`calcite-table >>> calcite-pagination >>> .${PAGINATION_CSS.page}[value="1"]`);
+      await numberLink.click();
+
+      expect(await table.getProperty("currentPage")).toBe(1);
+
+      table.setProperty("currentPage", 21);
+      await page.waitForChanges();
+
+      expect(await table.getProperty("currentPage")).toBe(3);
+
+      table.setProperty("currentPage", 0);
+      await page.waitForChanges();
+
+      expect(await table.getProperty("currentPage")).toBe(1);
+    });
+
+    it("starts on page out of upper range", async () => {
+      const page = await newE2EPage();
+      await page.setContent(
+        html`<calcite-table id="calcite-table" caption="Simple table" page-size="3" bordered current-page="100">
+          <calcite-table-row slot="table-header">
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 1</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 2</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 3</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 4</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+        </calcite-table>`,
+      );
+
+      const table = await page.find("calcite-table");
+
+      expect(await table.getProperty("currentPage")).toBe(2);
+    });
+
+    it("starts on page out of lower range", async () => {
+      const page = await newE2EPage();
+      await page.setContent(
+        html`<calcite-table id="calcite-table" caption="Simple table" page-size="3" bordered current-page="0">
+          <calcite-table-row slot="table-header">
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+            <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 1</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 2</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 3</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+          <calcite-table-row>
+            <calcite-table-cell>cell content 4</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+            <calcite-table-cell>cell content</calcite-table-cell>
+          </calcite-table-row>
+        </calcite-table>`,
+      );
+
+      const table = await page.find("calcite-table");
+
+      expect(await table.getProperty("currentPage")).toBe(1);
+    });
   });
 });