+++ packages/components/src/custom-theme/input-time-zone.ts
@@ -1,0 +1,9 @@
+import { html } from "../../support/formatting";
+
+export const inputTimeZoneTokens = {
+  calciteInputTimeZoneCornerRadius: "",
+};
+
+export const inputTimeZone = html`
+  <calcite-input-time-zone reference-date="2020-01-01" value="-60"></calcite-input-time-zone>
+`;