+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.0-next.16](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.15...@esri/calcite-components-react@5.2.0-next.16) (2026-07-02)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.15](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.14...@esri/calcite-components-react@5.2.0-next.15) (2026-07-01)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.14](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.13...@esri/calcite-components-react@5.2.0-next.14) (2026-07-01)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.12...@esri/calcite-components-react@5.2.0-next.13) (2026-07-01)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.11...@esri/calcite-components-react@5.2.0-next.12) (2026-06-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.10...@esri/calcite-components-react@5.2.0-next.11) (2026-06-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.2.0-next.10](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.9...@esri/calcite-components-react@5.2.0-next.10) (2026-06-30)
 
 **Note:** Version bump only for package @esri/calcite-components-react